Update README
First, make it an RST instead of markdown (needed for PyPI).
Then, update may of the details, notably the example model.

I'm not sure if it's worth keeping the OpenCL installation
instructions here; when I reinstalled OpenCL recently I just used
the PyOpenCL instructions, which are now linked, and they seemed
to be more up-to-date. I would recommend we just defer to them.