Store auto-created context in Simulator
Previously, we did this only for testing because tests made lots
of simulators, which created lots of contexts, which took up all
of the host memory. This commit handles this in the Simulator
itself so that users who are making a bunch of Simulators without
passing in a context don't run into the out of host memory error.

Also, some tests here should only be run on nengo_ocl.Simulator, so
there's no reason to use the Simulator fixture. As such, there doesn't
need to be a Simulator fixture defined in nengo_ocl, and instead we
should rely on the simulator being set in `tox.ini` when running Nengo
tests.