Fix kwarg handling in ast_conversion.
Not sure if this is a Python 3 change, but for me `expr` has no
'kwargs' attribute, so the assertion that it is None fails even
though the rest of the function should work fine since there are
no kwargs. Changed the assertion to also pass if there is no
kwargs attribute.

Also fixed printing the exception message in test_ast_conversion.
The exception message used to be stored as Exception.message, but
since Python 2.6 or so, the recommended way to print the messages
is to use Exception.__str__. In Python 3, Exception.message no longer
exists.