Fix future Nengo warning when using future Nengo
If the test suite is run with a future version of Nengo, it will
raise a warning about it at the start of the test suite. This ends
up then filtering out the warning that we're testing in
`test_warn_on_future_version`. Nengo instead uses monkey patching
rather than `pytest.warns`, so we use its implementation of `warns`.