Change step signal to integer type
This allows simulations to be run for many more timesteps.

This implementation only allows the step signal to be integer.
Ideally, a future improvement would group all signals by dtype
(there may be other integer signals). It would also allow signals
to be further split into separate buffers, so that all signals are
not in one large buffer, to better use card memory
(there is a limit that a buffer cannot use more than a quarter of
the card's memory).